feat: include fee bump in proof send modal (#186)
* feat: include fee bump in proof send modal

* feat: include submit proof modal in full user history table
