+++ web/assets/js/react/components/Modal/SubmitProof/SubmitStep.tsx
@@ -22,6 +22,8 @@ import { useToast } from "../../../state/toast";
 import { formatEther, toHex } from "viem";
 import { useChainId } from "wagmi";
 import { Button } from "../../Button";
+import { BumpFee } from "./BumpFee";
+import { fetchProofVerificationData } from "../../../utils/aligned";
 
 type Game = {
 	id: "beast" | string;
@@ -285,6 +287,62 @@ export const SubmitProofStep = ({
 		nonce,
 	]);
 
+	const [bumpLoading, setBumpLoading] = useState(false);
+	const formRetryRef = useRef<HTMLFormElement>(null);
+	const handleBump = async (chosenWei: bigint) => {
+		try {
+			setBumpLoading(true);
+
+			const res = await fetchProofVerificationData(
+				proofSubmission?.id || ""
+			);
+			if (!res) {
+				addToast({
+					title: "There was a problem while sending the proof",
+					desc: "Please try again.",
+					type: "error",
+				});
+				return;
+			}
+			const noncedVerificationData: NoncedVerificationdata =
+				res.verification_data;
+
+			noncedVerificationData.maxFee = toHex(chosenWei, { size: 32 });
+
+			const { r, s, v } = await signVerificationData(
+				noncedVerificationData,
+				payment_service_addr
+			);
+
+			const submitProofMessage: SubmitProof = {
+				verificationData: noncedVerificationData,
+				signature: {
+					r,
+					s,
+					v: Number(v),
+				},
+			};
+
+			setSubmitProofMessage(JSON.stringify(submitProofMessage));
+
+			addToast({
+				title: "Retrying submission",
+				desc: "Retrying proof submission using the newly selected fee.",
+				type: "success",
+			});
+
+			window.setTimeout(() => {
+				formRetryRef.current?.submit();
+			}, 1000);
+		} catch (error) {
+			addToast({
+				title: "Could not apply the bump",
+				desc: "Please try again in a few seconds.",
+				type: "error",
+			});
+		}
+	};
+
 	useEffect(() => {
 		if (
 			proofSubmission?.status === "pending" ||
@@ -318,6 +376,8 @@ export const SubmitProofStep = ({
 		}
 	}, [proofSubmission]);
 
+	const [bumpFeeOpen, setBumpFeeOpen] = useState(false);
+
 	if (
 		proofStatus === "pending" ||
 		proofStatus === "underpriced" ||
@@ -347,6 +407,57 @@ export const SubmitProofStep = ({
 					<p>Game: {game}</p>
 					<p>Level reached: {parsedPublicInputs?.level}</p>
 				</div>
+
+				{!bumpFeeOpen && (
+					<Button
+						className=""
+						variant="text-accent"
+						onClick={() => setBumpFeeOpen(prev => !prev)}
+					>
+						Bump fee
+					</Button>
+				)}
+
+				{(proofStatus === "pending" ||
+					proofStatus === "underpriced") && (
+					<form
+						ref={formRetryRef}
+						action="/proof/status/retry"
+						method="post"
+						className="hidden"
+					>
+						<input
+							type="hidden"
+							name="_csrf_token"
+							value={csrfToken}
+						/>
+						<input
+							type="hidden"
+							name="submit_proof_message"
+							value={submitProofMessage}
+						/>
+						<input
+							type="hidden"
+							name="proof_id"
+							value={proofSubmission?.id}
+						/>
+					</form>
+				)}
+
+				{(proofStatus === "pending" || proofStatus === "underpriced") &&
+					bumpFeeOpen && (
+						<BumpFee
+							onCancel={() => setBumpFeeOpen(false)}
+							onConfirm={handleBump}
+							isConfirmLoading={bumpLoading}
+							previousMaxFee={
+								proofSubmission?.submitted_max_fee || "0x0"
+							}
+							lastTimeSubmitted={
+								proofSubmission?.inserted_at || "0"
+							}
+						/>
+					)}
 			</div>
 		);
 	}
