+++ web/lib/zk_arcade_web/controllers/html/history.html.heex
@@ -50,6 +50,7 @@
                 payment_service_address={@payment_service_address}
                 user_address={@wallet}
                 explorer_url={@explorer_url}
+                batcher_url={@batcher_url}
             />
         </div>
     </div>
